requirements.txt
python-heatclient
python-keystoneclient
